refactor: MOve fastify server to store-indexer/bin/sqlite-indexer
