+++ packages/store-indexer/package.json
@@ -27,6 +27,7 @@
     "test": "tsc --noEmit --skipLibCheck"
   },
   "dependencies": {
+    "@fastify/cors": "^8.3.0",
     "@latticexyz/block-logs-stream": "workspace:*",
     "@latticexyz/common": "workspace:*",
     "@latticexyz/store": "workspace:*",
@@ -38,8 +39,10 @@
     "cors": "^2.8.5",
     "debug": "^4.3.4",
     "drizzle-orm": "^0.27.0",
+    "fastify": "^4.21.0",
     "rxjs": "7.5.5",
     "superjson": "^1.12.4",
+    "trpc-panel": "^1.3.4",
     "viem": "1.3.1",
     "zod": "^3.21.4"
   },
