+++ packages/store-sync/package.json
@@ -53,7 +53,6 @@
     "better-sqlite3": "^8.4.0",
     "debug": "^4.3.4",
     "drizzle-orm": "^0.27.0",
-    "fastify": "^4.21.0",
     "kysely": "^0.26.1",
     "rxjs": "7.5.5",
     "sql.js": "^1.8.0",
